requirements/base.txt
dj-database-url
Django
django-forms-bootstrap
django-tinymce
gunicorn
pytz
